# docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: Greenwich SmartFAQ API
  version: 0.1.0
servers:
  - url: https://smartfaq.university.local/api
    description: Production
  - url: http://localhost:8000
    description: Local development
paths:
  /auth/login:
    post:
      summary: Admin login (FR-001)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        '200':
          description: JWT issued
        '401':
          description: Invalid credentials or lockout
  /auth/me:
    get:
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Login:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
...